requirements.txt
alembic==1.16.4
annotated-types==0.7.0
anyio==4.9.0
betterproto==1.2.5
certifi==2025.4.26
cffi==1.17.1
click==8.2.1
cryptography==45.0.3
dotenv==0.9.9
ecdsa==0.19.1
fastapi==0.115.14
greenlet==3.2.3
grpcio==1.73.0
grpcio-tools==1.73.0
grpclib==0.4.8
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
multidict==6.4.4
mypy-protobuf==3.6.0
protobuf==6.31.1
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
Pygments==2.19.2
PyMySQL==1.1.1
pyOpenSSL==25.1.0
python-decouple==3.8
python-dotenv==1.1.0
python-jose==3.5.0
python-multipart==0.0.20
python-telegram-bot==22.2
rich==14.0.0
rsa==4.9.1
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
stringcase==1.2.0
typer==0.16.0
types-protobuf==6.30.2.20250516
typing-inspection==0.4.1
typing_extensions==4.14.0
uvicorn==0.35.0
websockets==15.0.1
